requirements.txt
Django==3.2.5
psycopg2==2.9.2 
uWSGI>=2.0.19.1,<2.1
Pillow==7.1.2
PyJWT==1.7.1
djangorestframework==3.13.1
django-cors-headers==3.5.0
django-mailer==2.1 
drf-yasg==1.21.4 
